refactor: Http 요청을 받고 처리하는 Dispatcher와 builder writer를 processor 패키지 하위로 변경
